Refactor state (#67)
